+++ api_handler.go
@@ -5,8 +5,13 @@ package hatchet
 import (
 	"encoding/json"
 	"fmt"
+	"log"
 	"net/http"
 	"strings"
+)
+
+const (
+	TOPN = 25
 )
 
 // apiHandler responds to API calls
@@ -20,13 +25,30 @@ func apiHandler(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusOK)
 	w.Header().Set("Content-Type", "application/json")
 	tokens := strings.Split(r.URL.Path[len(apiPrefix):], "/")
+	dbase, err := GetDatabase()
+	if err != nil {
+		json.NewEncoder(w).Encode(map[string]interface{}{"ok": 0, "error": err.Error()})
+		return
+	}
+	defer dbase.Close()
+	if dbase.GetVerbose() {
+		log.Println(r.URL.Path)
+	}
+
 	if len(tokens) == 3 {
 		tableName := tokens[0]
 		category := tokens[1]
 		attr := tokens[2]
 		if attr == "slowops" && category == "stats" {
 			orderBy := r.URL.Query().Get("orderBy")
-			b, err := getSlowOpsStats(tableName, orderBy)
+			if orderBy == "" {
+				orderBy = "avg_ms"
+			}
+			ops, err := dbase.GetSlowOps(tableName, orderBy, "DESC", false)
+			if err != nil {
+				json.NewEncoder(w).Encode(map[string]interface{}{"ok": 0, "error": err.Error()})
+			}
+			b, err := json.Marshal(ops)
 			if err != nil {
 				json.NewEncoder(w).Encode(map[string]interface{}{"ok": 0, "error": err.Error()})
 			} else {
@@ -35,7 +57,14 @@ func apiHandler(w http.ResponseWriter, r *http.Request) {
 			return
 		} else if attr == "slowops" && category == "logs" {
 			topN := ToInt(r.URL.Query().Get("topN"))
-			b, err := getSlowLogs(tableName, topN)
+			if topN == 0 {
+				topN = TOPN
+			}
+			logstrs, err := dbase.GetSlowestLogs(tableName, topN)
+			if err != nil {
+				json.NewEncoder(w).Encode(map[string]interface{}{"ok": 0, "error": err.Error()})
+			}
+			b, err := json.Marshal(logstrs)
 			if err != nil {
 				json.NewEncoder(w).Encode(map[string]interface{}{"ok": 0, "error": err.Error()})
 			} else {
@@ -51,7 +80,7 @@ func apiHandler(w http.ResponseWriter, r *http.Request) {
 		severity := r.URL.Query().Get("severity")
 		duration := r.URL.Query().Get("duration")
 		limit := r.URL.Query().Get("limit")
-		logs, err := getLogs(tableName, fmt.Sprintf("component=%v", component), fmt.Sprintf("limit=%v", limit),
+		logs, err := dbase.GetLogs(tableName, fmt.Sprintf("component=%v", component), fmt.Sprintf("limit=%v", limit),
 			fmt.Sprintf("context=%v", context), fmt.Sprintf("severity=%v", severity), fmt.Sprintf("duration=%v", duration))
 		if err != nil {
 			json.NewEncoder(w).Encode(map[string]interface{}{"ok": 0, "error": err.Error()})
@@ -66,15 +95,4 @@ func apiHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	json.NewEncoder(w).Encode(map[string]interface{}{"ok": 1, "message": "Hello Hatchet API!"})
-}
-
-func getSlowLogs(tableName string, topN int) ([]byte, error) {
-	if topN == 0 {
-		topN = 25
-	}
-	logstrs, err := getSlowestLogs(tableName, topN)
-	if err != nil {
-		return nil, err
-	}
-	return json.Marshal(logstrs)
 }
